🚚️ api: move route to src/enrolment context
