+++ api/lib/routes.js
@@ -12,7 +12,6 @@ import * as certificationLivretScolaire from './application/certification-livret
 import * as certificationPointOfContacts from './application/certification-point-of-contacts/index.js';
 import * as certifications from './application/certifications/index.js';
 import * as complementaryCertificationCourseResults from './application/complementary-certification-course-results/index.js';
-import * as countries from './application/countries/index.js';
 import * as frameworks from './application/frameworks/index.js';
 import * as healthcheck from './application/healthcheck/index.js';
 import * as lcms from './application/lcms/index.js';
@@ -46,7 +45,6 @@ const routes = [
   certificationIssueReports,
   certifications,
   complementaryCertificationCourseResults,
-  countries,
   healthcheck,
   lcms,
   memberships,
