+++ api/tests/certification/enrolment/unit/application/country-route_test.js
@@ -1,8 +1,8 @@
-import { countryController } from '../../../../lib/application/countries/country-controller.js';
-import * as moduleUnderTest from '../../../../lib/application/countries/index.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
+import { countryController } from '../../../../../src/certification/enrolment/application/country-controller.js';
+import * as moduleUnderTest from '../../../../../src/certification/enrolment/application/country-route.js';
+import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
-describe('Unit | Router | country-router', function () {
+describe('Certification | Enrolment | Unit | Application | country-route', function () {
   describe('GET /api/countries', function () {
     it('should exist', async function () {
       // given
